move to Core::Utils::Rails
