+++ lib/datadog/tracing/contrib/action_cable/integration.rb
@@ -3,7 +3,7 @@
 require_relative '../integration'
 require_relative 'configuration/settings'
 require_relative 'patcher'
-require_relative '../../../core/contrib/rails/utils'
+require_relative '../../../core/utils/rails'
 
 module Datadog
   module Tracing
@@ -33,7 +33,7 @@ module Datadog
           # enabled by rails integration so should only auto instrument
           # if detected that it is being used without rails
           def auto_instrument?
-            !Core::Contrib::Rails::Utils.railtie_supported?
+            !Core::Utils::Rails.railtie_supported?
           end
 
           def new_configuration
