+++ lib/datadog/tracing/contrib/active_record/integration.rb
@@ -7,7 +7,7 @@ require_relative 'patcher'
 require_relative '../component'
 require_relative '../integration'
 require_relative '../rails/ext'
-require_relative '../../../core/contrib/rails/utils'
+require_relative '../../../core/utils/rails'
 
 module Datadog
   module Tracing
@@ -37,7 +37,7 @@ module Datadog
           # enabled by rails integration so should only auto instrument
           # if detected that it is being used without rails
           def auto_instrument?
-            !Core::Contrib::Rails::Utils.railtie_supported?
+            !Core::Utils::Rails.railtie_supported?
           end
 
           def new_configuration
