--- sig/datadog/core/contrib/rails/utils.rbs
@@ -1,13 +1,0 @@
-module Datadog
-  module Core
-    module Contrib
-      module Rails
-        module Utils
-          def self.app_name: () -> String
-
-          def self.railtie_supported?: () -> bool
-        end
-      end
-    end
-  end
-end
