+++ sig/datadog/core/utils/rails.rbs
@@ -1,0 +1,11 @@
+module Datadog
+  module Core
+    module Utils
+      module Rails
+        def self.app_name: () -> String
+
+        def self.railtie_supported?: () -> bool
+      end
+    end
+  end
+end
