+++ spec/datadog/tracing/contrib/auto_instrument_examples.rb
@@ -4,7 +4,7 @@ RSpec.shared_examples 'rails sub-gem auto_instrument?' do
 
     context 'outside of a rails application' do
       before do
-        allow(Datadog::Core::Contrib::Rails::Utils).to receive(:railtie_supported?).and_return(false)
+        allow(Datadog::Core::Utils::Rails).to receive(:railtie_supported?).and_return(false)
       end
 
       it { is_expected.to be(true) }
@@ -12,7 +12,7 @@ RSpec.shared_examples 'rails sub-gem auto_instrument?' do
 
     context 'when within a rails application' do
       before do
-        allow(Datadog::Core::Contrib::Rails::Utils).to receive(:railtie_supported?).and_return(true)
+        allow(Datadog::Core::Utils::Rails).to receive(:railtie_supported?).and_return(true)
       end
 
       it { is_expected.to be(false) }
